fix: change search params to hash for sign message (according the spec)
